pyproject.toml
[project]
name = "ecommerce"
version = "0.0.0"
description = "ecommerce backend with Django app running in Docker."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "django==6.0",
  "celery==5.6.2",
  "django-debug-toolbar==6.1.0",
  "gunicorn==23.0.0",
  "psycopg==3.3.2",
  "redis==7.1.0",
  "ruff==0.14.10",
  "setuptools==80.9.0",
  "whitenoise==6.11.0",
  "pillow==12.1.0",
  "djangorestframework==3.16.1",
  "djangorestframework-stubs==3.16.7",
  "djangorestframework-simplejwt>=5.0,<6",
  "django-cors-headers>=3.13.0"
]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
extend-select = ["I", "SIM"]
